Hanin_CRUD_Interview_Question_Modification

+++ FrontEndAngular/src/app/employees/interviewer/interviews-list/interviews-list.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { InterviewsService } from '../../../Services/interviews.service';
 import { interviewList } from '../../../Models/interviews';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
+import { CreateInterviewComponent } from '../create-interview/create-interview.component';
 
 @Component({
   selector: 'app-interviews-list',
@@ -12,7 +14,7 @@ export class InterviewsListComponent implements OnInit {
 
   interviews! : interviewList[];
 
-  constructor(private interviewsService: InterviewsService , private router: Router ) {
+  constructor(private dialog :MatDialog, private interviewsService: InterviewsService , private router: Router ) {
    }
 
   ngOnInit(): void {
@@ -30,11 +32,14 @@ export class InterviewsListComponent implements OnInit {
   }
 
   deleteInterviews(id: number) {
+    let confirmation =confirm("Do you really want to delete this interview ?")
+    if(confirmation)
     this.interviewsService.deleteInterviews(id)
     .subscribe(o =>{
       this.interviews = o;
       this.loadList();
     });
+
 }
 interviewDetails(id: number){
   this.router.navigate(['employees/interviewer//detailinterview', id]);
@@ -43,5 +48,15 @@ updateInterviews(id: number){
   this.router.navigate(['employees/interviewer//updateinterview', id]);
 }
 
+refrech() {
+  window.location.reload();
+}
+
+onOpenDialogCreate():void{
+  const dialogConfig = new MatDialogConfig();
+  dialogConfig.disableClose = true;
+  dialogConfig.autoFocus = true;
+  this.dialog.open(CreateInterviewComponent, dialogConfig);
+}
 
 }
